New module plasma.nix; Additional WMs enabled, but no configs; Dotfiles cleanup.
Added sddm and plasma settings to plasma.nix in system modules.
displaymanager.nix remains for backup display-managers such as greetd, but currently not in use.
River, Sway, and Wayfire are now enabled for experimentation in hyprland-wm.nix and available in sddm on boot, however these are do not yet have config files added to home-manager, so they are currently clean-slate-state (read as 'opportunity' or 'unusable').
Also did some dotfiles and config cleanup."